I added data validation in RegisterActivity. I added three columns to ProductTable which enables storing info about amount of proteins, carbos and fat per 100g in a product. I also adjusted layouts in order to display those attributes

+++ app/src/main/java/com/example/dell/menu/screens/meals/addOrEdit/AddOrEditMealManager.java
@@ -280,7 +280,7 @@ public class AddOrEditMealManager {
             if(cursor.getCount() > 0){
                 cursor.moveToPosition(-1);
                 while (cursor.moveToNext()){
-                    result.add(new Product(cursor.getString(1), cursor.getInt(3), cursor.getString(2)));
+                    result.add(new Product(cursor.getString(1), cursor.getDouble(3), cursor.getString(2)));
                     result.get(result.size()-1).setProductId(cursor.getInt(0));
                 }
             }
